Move MntOps to lib.rs
In anticipation of mountinfo, move definitions that both will use to a
common location.
